ADS: Prevent saving not displayed workspace
Prevent saving a workspace that was never shown, e.g. only welcome page
was shown and QtDS was closed again. This scenario causes the workspace
to have wrong sizes hence the next time this workspace will be shown
it will look distorted.
This behavior is a result of the startup workspace always being loaded
and always saved. If the workspace was never visually shown, it didn't
get the correct geometry from the main window which results in wrong
sizes of splitters in the workspace.

Change-Id: I780b561e98e72f7d48becf32d178fb600b5d4336
Reviewed-by: Qt CI Patch Build Bot <ci_patchbuild_bot@qt.io>
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>