ExtensionSystem: Add plugin details dialog to error dialog
When a plugin fails to load, we show an error dialog with each of the
failed plugins and their error message. Add a button that opens the
plugin details dialog, and also show that on double click on a plugin in
the list.

Refactors opening a modal dialog with the details into its own function.

Change-Id: Id55d0add3682860470ede63b141fbce2e8dab75f
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>