EffectMaker: Bake preview shaders using correct version of qsb tool
If the target kit Qt version differs from Qt version that QDS was built
with, there is a chance that shaders baked for the target kit don't
work for rendering preview on QDS side, if QShader version has changed
between the two Qt versions. This means we must bake separate shaders
for the preview and the project to use, each using the correct version
of qsb tool.

Task-number: QDS-11361
Change-Id: I44df2aeb3f4d6aa4b80e3be052d46735d96c53ef
Reviewed-by: Qt CI Patch Build Bot <ci_patchbuild_bot@qt.io>
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>