Fonts: Add Medium SourceCodePro font variants
Add the Medium font families.

The Medium variants make sure that on a MacBook Pro with Retina display
the font doesn't look out of place being too thin.

Taken from https://github.com/adobe-fonts/source-code-pro/releases/tag/
2.030R-ro%2F1.050R-it

We need to stay with version 2.0.30 due to hinting on Windows, and to
force QFont::Medium on macOS due to the fact that the medium font is not
picked up. Newer versions like 2.0.42 do not have this issue.

Task-number: QTCREATORBUG-29964
Change-Id: I4c0fba5730c4a6b869d900642b675698c29c6ae8
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>