ManhattanStyle: Remove a hack for QComboBox wheel scrolling
Use Utils::setWheelScrollingWithoutFocusBlocked() explicitly
when a QComboBox is placed inside a scroll area.

The Utils::setWheelScrollingWithoutFocusBlocked() is more
general and may be applied to other widgets, like QSpinBox.

This patch brings back the wheel scrolling without a modifier key
for combo boxes placed in dialogs.

Keep the unconditional scroll wheel behavior when ctrl (meta on mac)
modifier is pressed.

Change-Id: Ieea0f228cea6b59e276bd3256eb6ea5ecc5b7d14
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>