Designer: Fix the mapping between buttons and edit mode
The tool buttons for switching the edit mode e.g. between "edit",
"signals/slots", and "buddy" hardcoded the index of the corresponding
tool in designer. The index in designer depends on the load order of
(static) plugins, that has changed in qtbase
2f226336a2aeb477b7ba339b6c3a63abb69746c3 (Qt 6.5.1 and later).

Instead of relying on the index/plugin load order, try to find the right
tool by class name (still falling back to the index otherwise).

It would be nicer if the tools had actual IDs though, but that change
would need to go into Designer first.

Fixes: QTCREATORBUG-30017
Change-Id: I17caea6289574db2b1c8934e80d865a142b74725
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>