Support view/hide right side bar in Debug mode
Like it is now supported in Widget Designer, the view/hide action
toggles the corresponding dock widget area. We cannot do that for the
left dock widget area yet, because that action still needs to toggle the
navigation side bar.

To support this, we need to save not only the QMainWindow state for the
debugger perspectives, but the whole FancyMainWindow state including the
information about hidden dock widget areas. Unfortunately the
PerspectiveState was serialized to QVariant without versioning, so
keeping it that way is difficult. Switch saving&restoring of
PerspectiveState to Utils::Store (and keep some compatibility code).

Change-Id: I0035841930c6a574aeb31650a28be9989e2b5741
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>