SquishTests: Attempt to stabilize designer test
The switch to cmake broke this test completely as qmake handles
changes to ui files differently.
Changing a ui file needs re-execution of moc which is not done
automatically for cmake except when building.
Beside this the cmake project manager seems to inform the code
model about respective changes significantly later, if at all.
So, we explicitly trigger a build now after changing the ui file
and explicitly ensure the code model treats the file correctly
by closing and re-opening it.

Change-Id: Icb9cfb572c1442eff1bebc9c29fdee55155ca287
Reviewed-by: Robert Löhning <robert.loehning@qt.io>