Doc: Fix configuration if qdoc is missing
As is currently the case for our Debian/ARM binaries.
It was failing when trying to make the html_docs_qtcreator target depend
on the doc_attributions target, because the html_docs target is not
created if qdoc is missing.

Amends 2ce9f9d97daa2dfe018e58a8b61af643add4b612

Change-Id: Ia5b118d6d373f7cc131d6fb9a03555180509b9d4
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Cristian Adam <cristian.adam@qt.io>