Welcome: Avoid duplicate examples when searching
When adding examples for a category/section we also add them to the list
of all examples. To avoid ownership issues we duplicate examples when
they appear in multiple categories, but that means that the existing
mechanism of filtering duplicates based on item pointer does no longer
work. Filter duplicates based on name+description.
It would possibly be nicer to give the examples IDs (serial numbers)
when reading them and just using that for filtering duplicates, but that
isn't possible in a binary compatible manner atm.

Fixes: QTCREATORBUG-30069
Change-Id: I0ee9ef7b86955af5ee8ccdb9c5683ced81097671
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>