ExternalTool: Fix output handling for General Messages
The handling of lines was wrong and only happened to work if the output
was delivered in one batch. Otherwise it would add line breaks - and if
the tool output was delivered not in batches of lines, that could break
in the middle of output lines.

Change to code to use the Process API for handling lines of output.
That also makes the codec handling in ExternalTool unnecessary. The
corresponding members can be removed in the next binary incompatible
release.

Fixes: QTCREATORBUG-29977
Change-Id: I9d37e5e374a43f34a28418272692cca17599c8b4
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>