PE: Make default build directory template relative to src dir
Qt Creator is using the following scheme to generate a build directory:
"../%{JS: Util.asciify("build-%{Project:Name}-%{Kit:FileSystemName}-%
{BuildConfig:Name}")}"

This implies that for docker devices the parent directory for the
project needs to mounted in order to have write access rights.

This is usually not the case and docker fails to configure projects.

Changing the template to "./build/%{JS: Util.asciify(%
{Kit:FileSystemName}-%{BuildConfig:Name}")}" makes sure that the build
directory is created
into a directory that has write access for docker.

Fixes: QTCREATORBUG-29972
Change-Id: I4839d76bf272ed7e59c60446636e4a394a0dea24
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>