Examples: Fix reading the category ordering from manifest
qtcreator_tutorials.xml contains ordering information for the tutorials,
and if we read that (first) we ignore any subsequent ordering that the
set of Qt examples defines (in qtdoc/examples-manifest.xml).

Only read the tutorials.xml if we are actually interested in tutorials.

Amends 08bbe885b49d6782178f0bc7b19c1bb162633666

Change-Id: I9148cc9f1060500f46be2bbc9c9ef6c2512c2f6c
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: Kai Köhne <kai.koehne@qt.io>