AutoTest: Fix location parsing for Qt Test
There are multiple ways to print the location depending on
the OS the test is executed on.
Do not "parse" again the location string, but use named
captures instead as they are present anyhow.
Former approach would return no location in case the location
would be printed like /path/to/file:10 which might be used on
UNIX.

Task-number: QTCREATORBUG-30143
Change-Id: If48bd0d9d9d8121522a44dfa69a15a0ccabde708
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: André Hartmann <aha_1980@gmx.de>