ProjectExplorer: Treat disabled UseLibraryPathsAspects as not selected
These are used to steer the addition of Qt library paths from
the build device to suitable environment entries of the run device
The aspect is hard-disabled for setups where these differ, but
so far checked only its value, which could be true independently.
Make the default access using operator()() also consider the
enabled state.

Change-Id: Iccf708999bad0336024c771e95a952118f2b2562
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>