iOS: Fix provisioning profile check for "newer" devices
Before deploying we check if the provisioning information includes the
device that is deployed to. This suffers from the same issue as
QTCREATORBUG-21291: We use the USB identifier for the device ID in Qt
Creator, which cannot contain dashes, but the iOS hardware udid that
Apple uses can contain dashes (and the USB identifier then is that udid
without the dashes).

We need to use the iOS hardware udid when checking for the device in the
provisioning information.

Fixes: QTCREATORBUG-30158
Change-Id: I6f9e162ad6ece02370923ea59c046f439b8cc47a
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>