Core: compress application icon updates
The current code updates the application icon for every task added to
the progressmanager. There are cases were we add a lot of tasks in one
go, for example when receiving diagnostics from the code model. This
might not even change the application icon (since it only updates for
errors), but we never the less create a new icon and send it to the os.

Optimization 1: Only trigger an update if the text differs from the
previous icon detail text.

Optimization 2: Do not directly update the icon but start a timer to
collect multiple updates and update the icon when the event loop is re-
entered.

This only affects Windows and Mac since the progressmanager_x11
implementation does not update the application icon.

Change-Id: I80cbb6f2520a83e79e2dba1b4079ab77fb7bf453
Reviewed-by: Eike Ziller <eike.ziller@qt.io>