TextEditor: avoid autoIndent when using default TextIndenter
Most programming languages use indentation to group stuff. So if we have
no special language support for a text file it is useful to start a
newline on the same indent depth as the previous line. But if we trigger
an explicit reindent (default ctrl+i) we usually don't want to change
the indentation if we don't know exactly where it should end up. This
also applies to refactoring changes, for example from the lsp.

Change-Id: I8268dc88cbceddfc7112757fd7d2fef940ab9a39
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>