TaskTreeRunner: Introduce aboutToStart(TaskTree *) signal
It enables the access to the internal TaskTree.
Emitted by TaskTreeRunner::start(), after the task tree
is instantiated and before it is started.

The lifetime of the passed task tree is controlled by the
TaskTreeRunner and is limited to the one particular execution.
Don't store pointer to the passed task tree.
If unvoidable, store the QPointer instead.

It may be used e.g. for connecting to the TaskProgress.

Change-Id: If0af11eb136c9cbff2e4c6e91f8b18a114880f74
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: hjk <hjk@qt.io>