Qt Creator redesign: Update the mode bar icons
This updates the existing mode bar icons in the svg file with the new Qt
Creator Design icons from Figma. The pngs are updated accordingly.

For consistency, the "beetle overlay" for toolbar actions is also
updated.

The custom color for "active" state of a mode icon has been removed. If
we keep this change, a couple of Theme::IconsMode*ActiveColor entries
can be removed.

Change-Id: I9b5422c79f5426e31b41e59d67ddc307b521c0af
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: hjk <hjk@qt.io>