CppEditor: avoid endless loop in symbol name lookup
Passing and empty text to QString::indexOf returns a 0 index and we
advance the start index by the length of the symbol name so we check
again from the start. Just assert if we pass an empty symbol name to
symbolOccurrencesInText and return an empty list of occurrences to avoid
this.

Task-number: QTCREATORBUG-30155
Change-Id: I6ad672a3dc2ae9d937fdc436874d07c00997fbf5
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>