LanguageClient: Fix npm installs for newer npm versions
We used the bin command line parameter for npm to get the path to the
runnable cmd inside the install location, but it seems this command is
missing for newer npm versions. Just check the default install location
and fall back to parse the npm list output in order to find the language
server path.

Change-Id: Ie144b061552c4a26fa6d54b232cef308669f7dbd
Reviewed-by: Artem Sokolovskii <artem.sokolovskii@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>