TextEditor: fix autoIndent
e70b99c7da49c27b519f2fcacb4e06974ef1f6ad broke the auto indent for all
indenters because those indenters are all based on the text indenter.

Add a fallback indenter that is used if no other indenter is configured
for a document, and avoid the auto indentation only if that fallback
indenter is used.

Change-Id: I00e04a07ab14c5fc433a71d1b9826f8e4cc5c0a0
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>