Allow linking of colors in the Palette section of a theme
This allows entries like toolbarBackgroundColor=background_Muted.
The code tries to resolve assignments until all colors are resolved
or in the last step no new color has been resolved.
Cycles are not resolved, but do not lead to issues.

In case some colors are not resolved we use the original code to assign
black and show an error message.

This also add the possibility to include ini files like this:
Includes="base.ini"

Change-Id: I31124a59dffca45d14b81d528ba7c88220e57577
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>