QmlDesigner: Close the header context menu on click
Since the context menu is defined inside the horizontal header view,
it didn't close by clicking on another column. The reason is that the
close policy works when we click outside the parent.
To resolve this, we can simply close the menu when the user clicks
on the header scope.
For right clicks, it's not mandatory to close the menu, since it will
be moved and opened again.

Fixes: QDS-11731
Change-Id: I5e0bb4c88dafdc54d0900a4c9e3bb584243373e4
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Shrief Gabr <shrief.gabr@qt.io>