Projects mode: Fix invisible tool bar above build system output
Amends d2500dc77b905a885d351367cf9154316ac73acf

For some reason, even if the title widget returns a size hint of height
0, it is still painted, but doesn't even reserve space in the layout.
So, the title was painted over the tool bar that we add to the dock
widget for the build system output in projects mode (in the right side
bar).

Set the visibility of the title widget in addition.

Change-Id: Ia00d73baba22294a57d689118565dc0fe3ee5617
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>