EffectComposer: Consider effect changes to be part of current document
This will allow saving effects when current document is saved for any
reason, including at application shutdown.

If effect is not yet saved once (i.e. it doesn't have a name yet), it
will not be saved in response to global save triggers to avoid
complications with requiring user input to provide the name.

Fixes: QDS-11436
Fixes: QDS-11446
Change-Id: I412ee4893e926d527b4f03f5f6c0c9b4e923bc1e
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>