Aspect: Don't try to change value more than once
When our window loses focus the editingFinished signal is triggered.
This can happen multiple times (QTBUG-121983) if the result of the edit
is a modal dialog and the user clicks outside our application.

This would call editingFinished again, resulting in multiple message
boxes being opened. (Easiest to reproduce is changing the build
directory of a project which displays a message box confirming the
change)

This patch checks, via a recursion Guard, that the slot is not already
in the process of being called.

Change-Id: I5e61e1b57680f0b1f2483a81682e791bfb5c0867
Reviewed-by: hjk <hjk@qt.io>