QmakeProjectManager: Fix Makefile compatibility check
E.g. qmake "binaries" for Android are actually shell scripts that call a
Desktop qmake in a different location, which lead Qt Creator to conclude
that the directory contains an incompatible build.
Fix this by checking that the -qtconf argument passed to qmake points
into the real qmake's parent directory, as is the case in the
abovementioned scenario.

Fixes: QTCREATORBUG-30354
Change-Id: Id6e878fab3379a3a8893389447514a1b7226784c
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>