SquishTests: Stabilize adding documentation
I noticed that Creator built on Qt 6.6.0 never needed longer than
two seconds to update the documentation after adding qtcreator.qch.
The same sources of Creator built on Qt 6.6.2 needed up to four
seconds. This subtle difference can lead to the situation that the
documentation was not yet updated when clicking "Getting Started"
in tst_WELP01, which then just displays an empty page.

Change-Id: I11ab2a902825190ed8ae27dad55a847f01ba44dc
Reviewed-by: Christian Stenger <christian.stenger@qt.io>