TaskTree: Implement asyncCount()
This might be useful for checking the integrality of tasks'
invocations and detecting the unwanted asynchronity in the task tree
execution.

Example: we may want to ensure that the setup handler of the 2nd task
was executed synchronously after the done handler of the 1st task,
without returning the control to the event loop.

This addresses the 34th point in the bugreport below.

Task-number: QTCREATORBUG-28741
Change-Id: I7213ee62a7555818cbe18fd7bb8362353a3be8ae
Reviewed-by: hjk <hjk@qt.io>