Axivion: Open issue information semi-automatic
This patch adds another way of opening issue information.
Double clicking an issue on the issues table will now
open the related information automatically beside jumping
to the respective location.
The original way of using a button on the tool tip of the
issue's text mark is clumsy, but still present.

Change-Id: I7b0fcafb8b01675fc020f8cc8174121ba7f52adc
Reviewed-by: hjk <hjk@qt.io>