Fix Fullscreen action text
On macOS the action was correctly switching between "Enter Full Screen"
and "Exit Full Screen", but on Windows & Linux the text was fixed to
"Full Screen" and the code tried to set the check state - but forgot to
make it checkable in the first place.

Actually it is unclear what the "correct" behavior is on Windows &
Linux. Neither on Gnome or KDE or Windows the action shows a check mark
when in Full Screen mode though. Either there is a tool button with an
icon, or some variation of "Exit Full Screen" or "Leave Full Screen".

Change the text of the action on all platforms, use "Exit Full Screen"
on all of them, but stay with just "Full Screen" on Windows & Linux (as
opposed to "Enter Full Screen" on macOS).

Fixes: QTCREATORBUG-30365
Change-Id: Ic55a30e32302ceb12f75449781b1aefecb370c97
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>