EffectComposer: Generate property specifics for effects
With proper specifics sheet, the effect will show the available
properties nicely in property view.

Also introduced a new concept of FooBarSpecificsDynamic.qml
to provide property specifics. The difference to regular
FooBarSpecifics.qml file is that the dynamic version will
be reloaded always, which allows us to change it at runtime.

Fixes: QDS-11995
Change-Id: I744124182428c5c4607856e7970700d9a5e9972e
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Qt CI Patch Build Bot <ci_patchbuild_bot@qt.io>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>