Axivion: Fix sort toggling glitches
The resize handle of a header has some margin to decide
whether to start a resize attempt.
This margin interfered with our randomly chosen one.
Take the needed margin into account and beside this store
an additional information about the column that triggered
our sort toggling attempt.
This avoids starting to toggle some column and moving
slightly to the next one and ending up in sorting the
final column of the move.

Change-Id: I8199f90fd06145c61d2c5e7574b6f040c72022b6
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>